+++ programs/drift/src/instructions/admin.rs
@@ -4960,6 +4960,50 @@ pub fn handle_update_delegate_user_gov_token_insurance_stake(
     Ok(())
 }
 
+pub fn handle_update_feature_bit_flags_builder_codes(
+    ctx: Context<HotAdminUpdateState>,
+    enable: bool,
+) -> Result<()> {
+    let state = &mut ctx.accounts.state;
+    if enable {
+        validate!(
+            ctx.accounts.admin.key().eq(&state.admin),
+            ErrorCode::DefaultError,
+            "Only state admin can enable feature bit flags"
+        )?;
+
+        msg!("Setting 3rd bit to 1, enabling builder codes");
+        state.feature_bit_flags = state.feature_bit_flags | (FeatureBitFlags::BuilderCodes as u8);
+    } else {
+        msg!("Setting 3rd bit to 0, disabling builder codes");
+        state.feature_bit_flags = state.feature_bit_flags & !(FeatureBitFlags::BuilderCodes as u8);
+    }
+    Ok(())
+}
+
+pub fn handle_update_feature_bit_flags_builder_referral(
+    ctx: Context<HotAdminUpdateState>,
+    enable: bool,
+) -> Result<()> {
+    let state = &mut ctx.accounts.state;
+    if enable {
+        validate!(
+            ctx.accounts.admin.key().eq(&state.admin),
+            ErrorCode::DefaultError,
+            "Only state admin can enable feature bit flags"
+        )?;
+
+        msg!("Setting 4th bit to 1, enabling builder referral");
+        state.feature_bit_flags =
+            state.feature_bit_flags | (FeatureBitFlags::BuilderReferral as u8);
+    } else {
+        msg!("Setting 4th bit to 0, disabling builder referral");
+        state.feature_bit_flags =
+            state.feature_bit_flags & !(FeatureBitFlags::BuilderReferral as u8);
+    }
+    Ok(())
+}
+
 #[derive(Accounts)]
 pub struct Initialize<'info> {
     #[account(mut)]