+++ sdk/src/index.ts
@@ -121,6 +121,7 @@ export * from './dlob/orderBookLevels';
 export * from './userMap/userMap';
 export * from './userMap/referrerMap';
 export * from './userMap/userStatsMap';
+export * from './userMap/revenueShareEscrowMap';
 export * from './userMap/userMapConfig';
 export * from './math/bankruptcy';
 export * from './orderSubscriber';