+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- program: builder codes ([#1805](https://github.com/drift-labs/protocol-v2/pull/1805))
+
 ### Fixes
 
 ### Breaking